Reset previous selection
When multi-select is called several times on the same 'select' HTML
element (e.g. when this 'select' element is updated with fresh data
from a database each time it is displayed), it is necessary to
reset previous selections (if any) before processing current
selected / unselected items, otherwise items that were in the class
'ms-selected' will remain in this class, and will therefore be
(wrongly) displayed as selected.