prop() instead of attr() for regular control
The proper way to select values in prop(), instead of attr(). Values which has been set up with attr() cannot be cleared!!
And this is big problem that cost more than 10 hours of my life today and before when I did not believe in this bug and tried many workarounds!